# [core] Improve performance of PartialUpdateMergeFunction with sequence group