<!-- Please specify the module before the PR name: [core] ... or [flink] ... -->

### Purpose

Improve the performance of PartialUpdateMergeFunction with sequence group.  

We made the following changes:

* Use a List, instead of a Map, to store the `FieldsComparator` and `FieldAggregator` to improve the performance for accessing the `FieldsComparator` and `FieldAggregator`.
* Get the field from the value row when it is needed to avoid the unnecessary heavy call of `FieldGetter#getFieldOrNull`.
* Use an array of booleans to record if the sequence group is empty to avoid duplicate checks for the same sequence group.

The optimization gives about 51% performance boost by the benchmark `PartialUpdateMergeFunctionBenchmark` with empty sequence groups, and 28% without empty sequence groups. According to our experience, the `PartialUpdateMergeFunction#updateWithSequenceGroup` method is about 60% workload of the end-to-end compaction with partial update merge engine and sequence group, so we can expect about 30% overall performance of such compaction.

The following benchmark results were run on

```
OpenJDK 64-Bit Server VM 1.8.0_412-b08 on Mac OS X 15.4
Apple M3 Pro
```

Before optimization
```
partial-update-benchmark:                                                                            Best/Avg Time(ms)    Row Rate(K/s)      Per Row(ns)   Relative
--------------------------------------------------------------------------------------------------------------------------------------------------------------------
OPERATORTEST_partial-update-benchmark_updateWithSequenceGroup                                           12079 / 12883           3311.4            302.0       1.0X
OPERATORTEST_partial-update-benchmark_retractWithSequenceGroup                                          10976 / 11129           3644.3            274.4       1.0X
OPERATORTEST_partial-update-benchmark_updateWithEmptySequenceGroup                                      11636 / 11727           3437.6            290.9       1.0X
```


After optimization
```
partial-update-benchmark:                                                                            Best/Avg Time(ms)    Row Rate(K/s)      Per Row(ns)   Relative
--------------------------------------------------------------------------------------------------------------------------------------------------------------------
OPERATORTEST_partial-update-benchmark_updateWithSequenceGroup                                            9995 / 10029           4002.0            249.9       1.0X
OPERATORTEST_partial-update-benchmark_retractWithSequenceGroup                                            9311 / 9330           4295.9            232.8       1.0X
OPERATORTEST_partial-update-benchmark_updateWithEmptySequenceGroup                                        7701 / 7747           5194.2            192.5       1.0X
```

### Tests

Already covered by the existing tests, such as `PartialUpdateMergeFunctionTest`